Feat: Render work cloud by @antv/g2
